+++ src/clj/rems/db/test_data_helpers.clj
@@ -217,5 +217,3 @@
                        :application-id app-id
                        :actor actor})
     app-id))
-
-