+++ src/cljs/rems/actions/review.cljs
@@ -29,13 +29,14 @@
                   :on-click #(rf/dispatch [::open-form])}])
 
 (defn review-view
-  [{:keys [application-id on-send]}]
+  [{:keys [application-id disabled on-send]}]
   [action-form-view action-form-id
    (text :t.actions/review)
    [[button-wrapper {:id "review-button"
                      :text (text :t.actions/review)
                      :class "btn-primary"
-                     :on-click on-send}]] ;; TODO disable submit if comment field is empty?
+                     :disabled disabled
+                     :on-click on-send}]]
    [:<>
     [comment-field {:field-key action-form-id
                     :label (text :t.form/add-comments-not-shown-to-applicant)}]
@@ -46,6 +47,7 @@
   (let [comment @(rf/subscribe [:rems.actions.components/comment action-form-id])
         attachments @(rf/subscribe [:rems.actions.components/attachments action-form-id])]
     [review-view {:application-id application-id
+                  :disabled (and (empty? comment) (empty? attachments))
                   :on-send #(rf/dispatch [::send-review {:application-id application-id
                                                          :comment comment
                                                          :attachments attachments